requirements.txt
# --- Core LangChain ---
langchain>=0.2.7
langchain-community>=0.2.6
langchain-openai>=0.1.8
langserve>=0.2.3
langchain-chroma

# -- (Opcional) En caso de fallar OpenAI ---
sentence-transformers

# --- Web server (FastAPI + Uvicorn) ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse_starlette

# --- Vector store ---
chromadb>=0.4.24

# --- Document Loaders ---
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-docx>=1.1.0
pypdf>=4.0.0

# --- Utilidades ---
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.7.0